+++ forge.config.ts
@@ -5,6 +5,7 @@ import type { ForgeConfig } from '@electron-forge/shared-types';
 import { MakerSquirrel } from '@electron-forge/maker-squirrel';
 import { MakerZIP } from '@electron-forge/maker-zip';
 import { MakerDeb } from '@electron-forge/maker-deb';
+import { MakerDMG } from '@electron-forge/maker-dmg';
 import { MakerRpm } from '@electron-forge/maker-rpm';
 import { WebpackPlugin } from '@electron-forge/plugin-webpack';
 
@@ -14,11 +15,19 @@ import { rendererConfig } from './webpack.renderer.config';
 const config: ForgeConfig = {
   packagerConfig: {
     icon: 'assets/icons/icon',
+
+    // This makes sure that the app is not shown in the dock on macOS.
+    extendInfo: {
+      LSUIElement: true,
+    },
   },
   rebuildConfig: {},
   makers: [
     new MakerSquirrel({}),
     new MakerZIP({}),
+    new MakerDMG({
+      // https://js.electronforge.io/interfaces/_electron_forge_maker_dmg.MakerDMGConfig.html
+    }),
     new MakerRpm({
       // https://js.electronforge.io/interfaces/_electron_forge_maker_rpm.InternalOptions.MakerRpmConfigOptions.html
       options: {
@@ -61,5 +70,26 @@ const config: ForgeConfig = {
     }),
   ],
 };
+
+// If the environment variable KANDO_OSX_SIGN is set, we sign the macOS app. This requires
+// certivicates to be installed on the build machine.
+if (process.env.KANDO_OSX_SIGN === 'true') {
+  if (config.packagerConfig) {
+    config.packagerConfig.osxSign = {};
+  }
+}
+
+// If the environment variable KANDO_OSX_NOTARIZE is set, we notarize the macOS app. This
+// requires your Apple Developer Account ID and an app specific password to be set for
+// your account.
+if (process.env.KANDO_OSX_NOTARIZE === 'true') {
+  if (config.packagerConfig) {
+    config.packagerConfig.osxNotarize = {
+      appleId: process.env.OSX_APP_SPECIFIC_ID || '',
+      appleIdPassword: process.env.OSX_APP_SPECIFIC_PASSWORD || '',
+      teamId: process.env.OSX_TEAM_ID || '',
+    };
+  }
+}
 
 export default config;