+++ webpack.ignores.ts
@@ -9,6 +9,10 @@ if (os.platform() !== 'win32') {
   ignores.push(/NativeWin32\.node$/);
 }
 
+if (os.platform() !== 'darwin') {
+  ignores.push(/NativeMacOS\.node$/);
+}
+
 if (os.platform() !== 'linux') {
   ignores.push(/NativeX11\.node$/);
   ignores.push(/NativeWLR\.node$/);